requirements.txt
pyyaml==5.4.1
robonomics-interface==1.6.1
discord.py==1.7.3